Some cleanups to use server side rendering and some simplifications
