+++ src/components/byteCollectionCategory/ByteCategoryEditor.tsx
@@ -12,10 +12,9 @@ import styled from 'styled-components';
 import SelectByteCollectionModal from './SelectByteCollectionModal';
 
 interface ByteCategoryEditorProps {
-  byteCategorySummary: CategoryWithByteCollection;
+  byteCategorySummary?: CategoryWithByteCollection;
   space: Space;
   viewByteCollectionsUrl: string;
-  upsertByteCategoryFn: (byteCollectionCategory: CategoryWithByteCollection) => Promise<void>;
 }
 
 const AddByteButton = styled.button`
@@ -32,7 +31,6 @@ function ByteCategoryEditor(props: ByteCategoryEditorProps) {
     space: props.space,
     viewByteCollectionsUrl: props.viewByteCollectionsUrl,
     byteCategory: props.byteCategorySummary,
-    upsertByteCollectionCategoryFn: props.upsertByteCategoryFn,
   });
 
   return (
@@ -57,10 +55,10 @@ function ByteCategoryEditor(props: ByteCategoryEditorProps) {
       <div className="my-4">
         <div className="flow-root">
           <ul role="list" className="-mb-8">
-            {byteCategory.byteCollectionArr!.map((byteCollection: ByteCollectionFragment, byteIndex: any) => (
+            {byteCategory.byteCollections.map((byteCollection: ByteCollectionFragment, byteIndex: any) => (
               <li key={byteCollection.id}>
                 <div className="relative pb-8">
-                  {byteIndex !== byteCategory.byteCollectionArr!.length - 1 ? (
+                  {byteIndex !== byteCategory.byteCollections.length - 1 ? (
                     <span className="absolute left-4 top-4 -ml-px h-full w-0.5 bg-gray-300" aria-hidden="true" />
                   ) : null}
                   <div className="relative flex space-x-3">
@@ -81,7 +79,7 @@ function ByteCategoryEditor(props: ByteCategoryEditorProps) {
                                 className="float-right ml-1"
                                 iconName={IconTypes.Trash}
                                 removeBorder
-                                disabled={byteCategory.byteCollectionArr.length === 1}
+                                disabled={byteCategory.byteCollections.length === 1}
                                 onClick={() => helperFunctions.removeByteCollection(byteCollection.id)}
                               />
                             </div>
